titanic-chatbot/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pandas==2.3.2
matplotlib==3.10.5
seaborn==0.13.2
python-dotenv==1.1.1
langchain==0.3.27
langchain-experimental==0.3.4
langchain-google-genai==2.1.9
langchain-huggingface==0.3.1
huggingface-hub==0.34.4
pytest==8.4.1
httpx==0.28.1
